requirements.txt
streamlit>=1.28.0
pathlib2>=2.3.7
requests>=2.31.0
python-dotenv>=1.0.0